refactor(#1361): extract channel_store_admin tests into module file
